Fixed the scale with 4 columns
